requirements.txt
requests
pandas
python-dotenv
streamlit
plotly
matplotlib
joblib